requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
dj-database-url==0.5.0
Django==4.1.4
django-adminlte3==0.1.6
django-cors-headers==3.13.0
django-database-url==1.0.3
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
psycopg2==2.9.5
python-dotenv==0.20.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
